Added tests for all get and post methods
