+++ client/src/__tests__/HealthOfficialPatientProfile.test.js
@@ -1,0 +1,6 @@
+import { render } from '@testing-library/react';
+import HealthOfficialPatientProfile from '../pages/HealthOfficialPatientProfile';
+
+test('Renders the DoctorViewingPatient page', () => {
+    render(<HealthOfficialPatientProfile />);
+  });
