+++ client/src/__tests__/HealthOfficialViewingPatient.test.js
@@ -1,0 +1,6 @@
+import { render } from '@testing-library/react';
+import HealthOfficialViewingPatient from '../pages/HealthOfficialViewingPatient';
+
+test('Renders the DoctorViewingPatient page', () => {
+    render(<HealthOfficialViewingPatient />);
+  });
