+++ package.json
@@ -14,7 +14,8 @@
   "license": "ISC",
   "devDependencies": {
     "cors": "^2.8.5",
-    "nodemon": "^2.0.15"
+    "nodemon": "^2.0.15",
+    "supertest": "^6.2.2"
   },
   "dependencies": {
     "@mui/icons-material": "^5.4.2",
